[BasicPruning] StoredNode and PruneCandidates encoder/decoders moved to encoding package
